refactor(validation): decompose pre_pr by check area into thin runner facade
Complete the pre_pr.py decomposition (issue #2223). Add the last area module
(checks_plugin.py: install-parity, plugin version-bump, hook-anchoring,
git-hooks, workflow-local-run) and rewrite pre_pr.py as a thin runner plus a
facade that re-exports every moved validator. The runner calls the same
validators in the same order with the same ADR-035 exit semantics; the diff is
a pure move with no behavior change.

pre_pr.py drops from 1442 to 484 lines, clearing the 500-line file-size
taste-lint that gates every PR. Each sibling check module is under 250 lines.

The validators now look up _run_subprocess and the base-ref helpers in their
own (bare-name) modules, consistent with the existing yaml_utils and
validate_design_review imports. Test mock targets move from
scripts.validation.pre_pr.* to the bare check-module names (checks_dash,
checks_plugin, checks_spec, checks_tooling, checks_common) so the mocks reach
the symbol where it is now looked up. No assertion changed.

Verification:
- tests/test_validation_pre_pr.py: 74 passed before and after.
- -k pre_pr across the suite: 75 passed before and after.
- python3 scripts/validation/pre_pr.py --quick --skip-tests: identical
  PASS/FAIL/SKIP set before and after (durations aside).

Fixes #2223
